+++ packages/abap-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/abap-deploy-config-inquirer
 
+## 1.1.10
+
+### Patch Changes
+
+-   Updated dependencies [0c64478]
+    -   @sap-ux/inquirer-common@0.5.11
+
 ## 1.1.9
 
 ### Patch Changes